presto-admin: Refactor directory structure using cookiecutter
Summary:
Using a cookiecutter template (https://github.com/audreyr/cookiecutter-pypackage),
generated a new package structure, including blank Sphinx docs, a test folder,
and setup.py/wheel for deploying Python.

We also switch from using an egg file for Fabric to adding it as a requirement --
aka a package that must be installed in the virtualenv in order to run the program.
This is the standard way to handle dependencies in Python, from what I've seen.
This installation is handled automatically by a combination of the setup.py script
and the requirements.txt file.

The development process is as follows:
1. Assuming you have virtualenvwrapper installed:
```
    $ mkvirtualenv prestoadmin
    $ python setup.py develop
```
This installs all of the dependencies listed in requirements.txt.

2. In order to test changes after they're made:
```
    $ flake8 prestoadmin tests
    $ python setup.py test
    $ tox
```
We will enforce flake8 Python style guidelines, with a few exceptions that
are listed in setup.cfg.  The changes made to the .py files are almost all
in order to comply with flake8.

3. In order to exit out of the virtualenv after finishing development for
this project:
```
    $ deactivate
```

4. In order to build the Sphinx documentation,
```
    $ make clean docs
```

Task: SWARM-100
Review Url: @@review_url@@

Test Plan: bin/presto-admin, flake8, python setup.py test, make clean docs

Reviewers: rs186079, as186127

Reviewed By: rs186079, as186127

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D16