presto-admin: Update version so that it works with setup.py
Summary:
Python doesn't have a good way of handling the version of a piece of
software.  The recommended method is to have the version both in
setup.py and in the source code. There are also a few other methods:
https://python-packaging-user-guide.readthedocs.org/en/latest/single_source_version.html
But none of them are that great.

Task: SWARM-165
Review Url: @@review_url@@

Test Plan: python setup.py develop in new virtualenv, tox, setup.py test

Reviewers: mf186042, rs186079

Reviewed By: rs186079

Subscribers: as186127

Differential Revision: https://phabricator.td.teradata.com/D90