presto-admin: Read topology in main, so we can use roles
Summary:
In order to use the @worker and @coordinator roles, we need
to fill in the roledef list in main, so that the mapping is
in place before running the Fabric tasks.

Additionally, we need to add all of the hosts to env.hosts,
so that Fabric knows about them.

Task: SWARM-59
Review Url: @@review_url@@

Test Plan: manual

Reviewers: as186127, rs186079

Reviewed By: as186127, rs186079

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D100