presto-admin: Topology show uses state.env
Summary:
Now that we read the topology config file in main.py, we want to use the
Fabric internal state to get the coordinator and worker nodes.

This patch does remove the try/catch for the ConfigurationError.  That
will be added in a separate patch that adds a decorator that checks the
configuration error that has been added to state.env while the topology
configuration was being read in.

Task: SWARM-59
Review Url: @@review_url@@

Test Plan: tox, lint, coverage

Reviewers: as186127, rs186079

Reviewed By: rs186079

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D107