Basic packaging of presto admin tool
Summary:
Added bdist_prestoadmin setuptools command that builds the wheel for prestoadmin, retrieves a tarball of virtualenv, generates an install script, and creates a tarball of everything. To build this, simply run: python setup.py bdist_prestoadmin.

The workflow to install is:
1. extract the tar ball
2. run the install-prestoadmin.sh script

The install script will extract the virtualenv tarball, create a virtual environment and pip installs the wheel for prestoadmin into it.

In a future patch, the tarball + script workflow will be replaced with an rpm.
Currently this process assumes connectivity to the internet.

Task: SWARM-166, SWARM-232

Test Plan: unit tests in test_bdist_prestoadmin

Differential Revision: https://phabricator.td.teradata.com/D120