presto-admin: Fix order of precedence of env variables
Summary:
We were not correctly handling the order of precendence of env variables,
so all of the config.json options (including hosts) were being wiped out.
In order to fix this, we perform the parsing twice, the first time just
saving the options that changed and the second time allowing the default
options as well.  Then, you first set all of the default options, then
the config.json options, then the options set on the CLI.

The order of precedence of env variables is:
1. options set on the CLI (e.g. --hosts, -user)
2. options set in config.json (e.g. the presto-admin topology)
3. default options that Fabric has

Task: SWARM-59
Review Url: @@review_url@@

Test Plan: tox, flake8, coverage, manual

Reviewers: as186127, rs186079

Reviewed By: as186127, rs186079

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D116