Read in and validate presto coordinator configuration
Summary:
This adds the ability to read in and validating the coordinator
configuration, and writing it to temporary local files.  This is the
first step in being able to deploy the configuration on the cluster.
Still to do: don't error out if coordinator.json does not exist

Task: SWARM-204

Test Plan: unit tests

Reviewers: as186127, cwallin

Reviewed By: cwallin

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D110