presto-admin: Add initial logging framework
Summary:
In this initial logging framework, we move over some logging/application
utilities from hadapt-admin.  In particular, we use the @entry_point
decorator, which wraps the main method and indicates the application
name, version, and log file path.  It also handles exceptions and redirects
those that float up to the top level to the log.

In order to use the @entry_point decorator, we had to create the
presto-admin script via setup.py, because bin/presto-admin launches
the python program in a non-standard way.  In order to generate this
script, just run python setup.py develop again.

Currently, most errors are swallowed/handled separately by Fabric, though,
so, the presto-admin log currently only contains the initial arguments
to presto-admin, the logging config file location, and the fact that the
program exited successfully.  As an example, I also added simple logging
for topology show.  The next tasks are to a) add audit logging
for the rest of the tasks we have so far and b) alter how Fabric catches
exceptions so as to be able to log them.

Task: SWARM-205
Review Url: @@review_url@@

Test Plan: manual testing of log file, lint, test-all

Reviewers: rs186079, as186127

Reviewed By: as186127

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D131