presto-admin: Pass in the proper hosts and excluded hosts
Summary:
This commit passes in explicitly the hosts and excluded hosts on
a per-task basis.

The Fabric @role annotation overrides pass in hosts via --hosts, but
passing in per-task hosts overrides @role.  Ideally, what we want
to do is pass in --hosts and for the workers() task, run the task on
all of those hosts that are workers.  Unfortunately, Fabric doesn't
do that for you, so we have to simulate that manually by checking
if a given host is a coordinator or a worker.

Additionally, we add wheel as a testing dependency because packaging
needs it (and make coverage was failing otherwise).

Task: SWARM-227
Review Url: @@review_url@@

Test Plan: manual on slippery and unit tests

Reviewers: rs186079, anu

Reviewed By: rs186079, anu

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D157