presto-admin: Ensure all exceptions are printed to the logs
Summary:
This commit contains an assortment of customer facing exceptions/logging
changes.

* Removes --show and --hide options for controlling Fabric output levels
* hides 'running' fabric status
* adds logging instead of print statements in main.py
* adds FabricApplication that does all of the exception handling and
disconnects the connections, so that main() is much simpler
* fixes a bug in Appliction such that exit("error message") shows
up in the logs. This style of exits is used very often by Fabric
for internal errors.

Still remaining for logging:
* The error messages for optparse are only logged to the CLI, not to the
logs. In order to fix this, we need to make a subclass of optparse.

Task: SWARM-205
Review Url: @@review_url@@

Test Plan: lint, test-all

Reviewers: rs186079, anu

Reviewed By: rs186079, anu

Subscribers: mf186042, an186016

Differential Revision: https://phabricator.td.teradata.com/D147