Add support for presto-admin connector add
Summary:
This adds the ability to add a connector.  To add a connector, you put
the configuration in connectors.json. For example
{"tpch.properties":{"connector.name":"tpch"}}
To add the connector, run
presto-admin connector add tpch
This will deploy the tpch configuration on all nodes in the cluster
If no name is specified, all connectors in connectors.json will be
added.

Test Plan: unit tests, manual test

Reviewers: anu, cwallin

Reviewed By: anu, cwallin

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D175