packaging: Include the logging .ini file in the source distribution
Summary:
The source distribution (created by sdist) relies on the MANIFEST.in
to indicate which files to include, not the package files specified in
setup.py.  With this change, the default logging file is included in
the distribution.

Task: SWARM-279
Review Url: @@review_url@@

Test Plan: manually verified .ini files exist in tar

Reviewers: anu, mf186042

Reviewed By: anu, mf186042

Subscribers: rs186079

Differential Revision: https://phabricator.td.teradata.com/D193