packaging: Change tar.bz2 to be owned by root
Summary:
Tar has pretty weird default behavior, in that by default, when it's
run by a normal user, it keeps that user as the owner. If it's run
by root, it looks at the archived directory to get its owner. Since
I generated the archive, the owner was cwallin, which was not recognized
on a remote host.  Thus, tar set the owner to 1000.

The different owner (in addition to different permissions) caused
SSH authentication issues when this tarball was untared in /root.

Because of the behavior of tar, it doesn't hurt to set the owner and
group as root: in the case in which it's run by a normal user, the
owner is changed to that user. And if it's run by root, it stays root.

The way to untar while not changing the surrounding directories is
the following:
tar xvf prestoadmin-0.1.0.tar.bz2 --no-overwrite-dir

This commit also includes some lint fixes.

Task: SWARM-283
Review Url: @@review_url@@

Test Plan: manual

Reviewers: mf186042

Reviewed By: mf186042

Subscribers: rs186079, anu

Differential Revision: https://phabricator.td.teradata.com/D185