presto-admin: Add configuration for where the connectors are
Summary:
The directory where catalog configs live, specified by plugin.config-dir,
is in /etc/presto/catalog. That's different than the location that Presto
expects it to be, so the server doesn't pick up any catalog configs if
you don't change that var.  Similarly, you need to change the location
of the plugin dir.

We put these properties in node.properties because they apply to a single
installed instance of Presto on a node (which is what node.properties is).
config.properties, on the other hand, is the configuration for the server.

Task: SWARM-282
Review Url: @@review_url@@

Test Plan: make clean lint test-all, manual

Reviewers: rs186079, anu

Reviewed By: anu

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D196