parser: Display detailed command info when args are wrong
Summary:
When a user types in the wrong number of arguments to a command, we
want to give a reasonable and informative error message.

This commit
* Changes display_command() to pass in the exit code, since we want an
exit code of 2 for a parsing error
* Adds an except block for TypeError, and then further checks to see whether
the exception is for arguments to a function, so we don't let the wrong
errors through to the user
* removes suppress_errors logic because it wasn't fully implemented
anyway
* adds testing -- all of the @runs_once commands need to have the
return value cleared before or after each run.

Task: SWARM-229
Review Url: @@review_url@@

Test Plan: make clean lint test-all coverage, manual

Reviewers: anu, rschlussel

Reviewed By: anu, rschlussel

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D227