presto-admin: package dependencies in the presto-admin tarball
Summary:
We want to package dependencies in the presto-admin tarball so that one can install presto-admin without a network connection.

python setup.py bdist_prestoadmin will download the dependencies, and package them in a third-party directory to be include in the tarball. The install script will direct pip to look at this directory when install the dependency requirements.

Will later need to add system tests when the framework is ready.

Task: SWARM-236

Test Plan: test_bdist_prestoadmin

Differential Revision: https://phabricator.td.teradata.com/D241