parser: Add extended help text
Summary:
Fabric has a lot of options that we want advanced users of presto-admin
to be able to use. However, we want to show those in an extended help
text because there are a lot that don't apply to the typical user.

In order to do that, we had to extend optparse. So, we
* add a new option for --extended-help
* add a HiddenOptionGroup class that can suppress the help text
for a whole OptionGroup, not just an Option
* add a print_extended_help method to LoggingOptionParser which
shows all of the help text for all of its option groups, if they
are HiddenOptionGroups, else just prints the help as usual.

Task: SWARM-239
Review Url: @@review_url@@

Test Plan: make clean lint test-all coverage, manual

Reviewers: rschlussel, anu

Reviewed By: anu

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D225