Fix issues with building sphinx docs for presto-admin
Summary:
Needed to add dependencies to requirements.txt for spinx.
Also added missing dep for flake8

Also modified Makefile slightly to not try to open the docs at the end. Automation would not needs this. However, there is a new Make target if one want to open the docs that way.

Test Plan: TODO when we have automation

Reviewers: cwallin

Reviewed By: cwallin

Differential Revision: https://phabricator.td.teradata.com/D298