Implement configuration show
Summary:
configuration show [node|jvm|config] prints the corresponding
configuration files to the user. If the file is missing we print a warning
and if no arguments are given we print all three configs

We rename the existing configure task to configuration deploy
[all|workers|coordinator] and hence rename the existing configuration.py
to configure.py to avoid conflicts with the fabric task name.

Test Plan: unit test, manual

Reviewers: cwallin, rschlussel

Reviewed By: rschlussel

Subscribers: an186016, mf186042

Differential Revision: https://phabricator.td.teradata.com/D296