presto-admin: Add ability to set arbitrary env variables on the CLI
Summary:
For Jenkins, we want to use dzdo instead of sudo in Fabric. Thus,
we need to be able to set the sudo command.

This ability was already in hfab/fabric's main.py, we just removed it so
as to simplify the code (and now are adding it back).

Task: SWARM-362
Review Url: @@review_url@@

Test Plan: make clean test-all coverage

Reviewers: rschlussel, anu

Reviewed By: anu

Subscribers: mf186042, an186016

Differential Revision: https://phabricator.td.teradata.com/D358