presto-admin: Try alternative temporary dir for fabric put
Summary:
Fabric's default temp dir for put is ~/. However, on Jenkins, root doesn't
have write access to ~/.  Thus, it fails.

We add fallback in this case to use /tmp, in case customer environments are
similar to Jenkins.  It's not fault-proof -- /tmp can also not have write
perms -- but it's worth a shot.

Eventually we want to make the REMOTE_PACKAGES_PATH configurable, so
as to avoid this sort of issue.

Task: SWARM-362
Review Url: @@review_url@@

Test Plan: make clean lint test-all

Reviewers: anu, rschlussel

Reviewed By: rschlussel

Subscribers: mf186042

Differential Revision: https://phabricator.td.teradata.com/D368