Fix server status to be more reliable
We were running server status once on the coordinator to get the entire
cluster's status, meaning, if the coordinator was down then the
server status failed too.

To avoid this, we now connect to each node and run the
system.runtime.nodes query. So each node does its own status check.

But since every node returns the entire cluster node info, we
additionally add a WHERE clause to the pushed sytem.runtime.nodes query, to get the
single result row for that particular node:
SELECT ... FROM sytem.runtime.nodes WHERE url_extract_host(http_uri) =
external.ip.of.the.pinged.node.

The external ip of the node is obtained from parsing http_uri column
from the system.nodes table. The node.id(uuidgen) from the node.properties
file on each node is taken and the system.nodes table is queried to get the
corresponding http_uri column (which is parsed to get the external
ip).

If server is down on a node, then we do not query that node and report
that no information is available.

Server Status:
	Node1(IP: IP1 roles: coordinator, worker): Running
	Node URI(http): http://active/statement
	Presto Version: presto-main:0.97-SNAPSHOT
	Node is active: True
	Connectors:     hive, system, tpch
Server Status:
	Node2(IP: IP2 roles: worker): Not Running
	No information available

Task: SWARM-439