presto-admin: Add helper script to install docker, and check if docker is installed
Summary:
We add a script to install docker on Ubuntu 14.04, and add a check in the
tests if docker is installed.  We don't want to install docker during the
tests because you don't want to run sudo during the tests. But regardless, we
need a better error message if docker isn't there.

Task: SWARM-451
Review Url: @@review_url@@

Test Plan: ran with Docker not installed and got error; ran install script manually

Reviewers: rschlussel

Reviewed By: rschlussel

Subscribers: mf186042, anu

Differential Revision: https://phabricator.td.teradata.com/D433