presto-admin: Default to warn_only when running commands
Summary:
We want to default to warn_only when running commands, because on a
distributed system, it is expected that some nodes may fail.  The failure
of those nodes does not mean that the command itself failed.

We need to monkey patch Fabric to add the ability to log stdout/stderr
from the remote console (for sudo() and run()) and to add the
ability to log the message being passed to warn().

We also need to add a --fail-on-error flag, so that you can
turn off --warn-only.

Once we have warn_only by default, we can't catch an exception
from a failing put in package.py. Instead, we need to check the
return value of put.

Additionally, we can handle connector.remove properly and not have to
use a try/except.

Finally, updated some of the logging levels from debug to info.

Task: SWARM-343
Review Url: @@review_url@@

Test Plan: make clean lint test-all coverage; manual

Reviewers: rschlussel, anu

Reviewed By: rschlussel, anu

Subscribers: an186016, mf186042

Differential Revision: https://phabricator.td.teradata.com/D425