Fix issue with configurations overwriting each other and add product test for install
Summary:
This patch fixes the issue where coordinator configuration wasn't
getting set correctly it had the same temp location as workers.

It adds a product test for presto-admin server install that verifies
installation and configuration.

It fixes an issue with passwordless ssh on the cluster

Task: SWARM-366

Test Plan: ran the test

Reviewers: anu, cwallin

Reviewed By: anu, cwallin

Differential Revision: https://phabricator.td.teradata.com/D452