presto-admin: Detect correct port for prestoclient
Summary:
Before, we just used 8080 as the default port all the time. However,
a user can change the port in their configuration files.

Thus, we detect the port from the config files.

We also refactor the status/prestoclient logic so that less new
PrestoClients are created. We need to clear out the old rows
when we do that.

We also add a system test for the port not being 8080, and refactor
some of the existing product tests to keep it DRY.

Task: SWARM-455
Review Url: @@review_url@@

Test Plan: manual, unit tests, system test

Reviewers: anu, rschlussel

Reviewed By: anu, rschlussel

Subscribers: mf186042, br186017

Differential Revision: https://phabricator.td.teradata.com/D467