presto-admin: update directory names to have no dash
Summary:
If you untar in /opt, you'll get the directory /opt/prestoadmin. We
were creating an /opt/presto-admin/packages directory, which was
confusing. We change all of the directories that are presto-admin to
prestoadmin.

There are a couple of changes needed in the Jenkins scripts, which
will go in after this does.

Test Plan: make lint test system-test

Reviewers: anu, rschlussel

Reviewed By: rschlussel

Subscribers: mf186042, br186017

Differential Revision: https://phabricator.td.teradata.com/D500