exceptions: Make ConfigurationError a UserVisibleError
Summary:
In this commit, we make ConfigurationError a UserVisibleError. We also move
ConfigurationError to utils/exception.py, so that all of the exceptions
are in the same place.

We want it to be a UserVisibleError because it means that the error message
is displayed to the user.

Task: SWARM-374
Review Url: @@review_url@@

Test Plan: make clean lint test-all

Reviewers: rschlussel, anu, anton

Reviewed By: anu, anton

Subscribers: anton, mf186042

Differential Revision: https://phabricator.td.teradata.com/D509