Fix issue with connector directory exists
Summary:
There was a race condition in multiple threads trying to create a
connectors directory at the same time.  This adds a try catch instead of
the if statement check to avoid the race condition.  It also only
creates the file if it does not already exist

Test Plan: make test-all

Reviewers: cwallin, anton

Reviewed By: cwallin, anton

Subscribers: mf186042, cwallin, anu

Differential Revision: https://phabricator.td.teradata.com/D512