main: Skip bad hosts by default
Summary:
We don't want presto-admin to abort if it can't reach one of the
hosts (which is the default Fabric behavior).  So, we set
skip_bad_hosts to True, and remove the CLI option since it's already
always true.

We don't add an option for the opposite of skip_bad_hosts -- it doesn't
seem as necessary, and if it becomes something that customers want
to do, they can always set it manually with the "set" command.

Task: SWARM-499
Review Url: @@review_url@@

Test Plan: make clean lint test-all; forthcoming product test

Reviewers: anu, rschlussel

Reviewed By: anu, rschlussel

Subscribers: anton, akshat, mf186042

Differential Revision: https://phabricator.td.teradata.com/D533