presto-admin-tests: Additional authorization tests
Summary:
We additionally test
* Specifying a keyfile with -i
* running presto-admin not as sudo
* SSH and sudo passwords being different

We change the behavior such that if presto-admin is not run as sudo, it fails
fast.

We do not write tests for interactive password input, because doing so is
difficult since getpass explicitly does not read from stdin, only TTY.
Additionally, it's pretty involved to set up CentOS such that the sudo
password is different than the SSH password, so we leave that for later.
We will just document that the SSH password and sudo password must be
the same.

Task: SWARM-499
Review Url: @@review_url@@

Test Plan: make clean lint test-all

Reviewers: anu, rschlussel

Reviewed By: anu, rschlussel

Subscribers: akshat, anton, mf186042

Differential Revision: https://phabricator.td.teradata.com/D571