pa-tests: Don't sort output from lost_connection test
Summary:
You can't sort the output from the lost connection test because it
expects one of two error messages.  This changes the test to instead
look for particular error messages.  You also can't remove the text that
was within the stacktrace because interleaved is sometimes important
messages.  Instead just ignore the stacktrace for comparison purposed
but leave it in the output string. Fixes automation issues.

Test Plan: ran in tox

Reviewers: anton, cwallin, anu

Reviewed By: anu

Subscribers: anu

Differential Revision: https://phabricator.td.teradata.com/D609