infra: Save snapshot of Docker container with Presto installed
Summary:
The tests take a long time to run (40+ min), primarily because it takes
between 1 and 2 minutes (sometimes longer) to install Presto.

Thus, we do a docker commit of the containers with default
Presto installed.  There is no mechanism to clean after a
whole suite of tests in unittest, so instead we clean up
those docker images in clean-test, as well as before you
make the bdist_prestoadmin in BaseProductCase.  Presumably,
you only want your Presto images to change when your Presto
dist changes, so it's ok for them to stick around in the
meantime.

For the start/stop/restart tests, this patch decreases the
runtime from 1063.702s to 841.369s.

Task: SWARM-276
Review Url: @@review_url@@

Test Plan: make lint test-all

Reviewers: anton, rschlussel

Reviewed By: anton, rschlussel

Subscribers: mf186042, anu

Differential Revision: https://phabricator.td.teradata.com/D595