Error handling: Abort current task on failure
Change behavior to abort the running task for a particular host on
failure, so subsequent tasks don't get run.  Continue with tasks for
other hosts.  This means that if there is an issue with one node, other
nodes are not affected, but the task on that particular node is stopped.

Testing: make test-all