presto-admin: make presto-admin parse teradata versions
Now we are checking that presto is installed before checking version using 'rpm -q presto' and 'rpm -q presto-server-rpm'.
So, now we will get 'Presto is not installed' message just in case when presto is not installed at all.

Then if presto is installed, we are trying to parse presto version usin '(\d+)\.(\d+)t?([-\.]SNAPSHOT)?' regexp.
First group - is the major version, second group is the minor version.
This regexp should parse all possible versions including teradata ones and snapshot ones.

When version doesn't match this pattern we are showing to user 'Incorrect presto version' message
instead of telling him that presto is not installed at all.

Testing: manual