Create new docker client for each call
When communicating with docker over http (this is the case when
boot2docker is in use) command could fail if multiple streaming
commands were executed using same docker client.
The request then failed with "Hijack is incompatible with use of
CloseNotifier" error.
Workaround for that (until docker-py library is fixed) is to create
new docker client everytime we need one.