Archive generated by bdist_prestoadmin should not include .
This change adds a base_dir argument to make_archive so that the archives we
generate do not include an entry for the current directory (.). Without this
change, we apply the permissions of some arbitrary directory on the machine
where the archive is built to the parent directory of prestoadmin when we
untar the archive. This is bad.