Use the dummy presto rpm in product tests when possible
The product tests install presto a large number of times to test various
behaviors of the installation code. In most cases we don't need to actually
install presto, because we don't do anything that requires having presto
installed (e.g. starting presto). For these cases, we can install the dummy
presto rpm instead. Because it's much smaller, presto-admin spends less
time sending the rpm to the nodes in the cluster, and the tests run faster.