Refactor presto/prestoadmin installation out of base_product_case.
In order to support writing product tests for the YARN integration with
slider, we need to refactor the product tests so that we can reuse the
existing docker/configurable cluster logic with slider-based presto
installation and configuration. This change extracts the installation of
presto-admin, presto, and setting the default topology from base_product_case
and moves it to individual modules.