Improve clean-test and clean-test-containers targets
clean-test tries to clean up all of the docker images that the product tests
create, and if it fails, it should show a helpful message telling the user how
to kill off any running containers. Add a dash to the front of the image
cleanup line so that make ignores the return code and actually shows you the
help if docker couldn't delete an image.

clean-test-containers deletes any running containers that got left behind by
the product tests. Piping a list of containers in to xargs docker kill causes
and error if there aren't any running containers, so I switched it to a loop
like the line that handles deleting images.