Make stdout/stderr capture optional in BaseTestCase
In order to debug issues in the product tests that we only see on OpenStack,
it would be super helpful to be able to print messages for debugging and see
the output, as we can't log in to any of the docker containers or even debug
the tests themselves. This change makes output capture optional and enables it
for the unit tests that require it.