Exit with a non-zero exit code when operations fail.
presto-admin should adhere to the convention that a failure results in a
non-zero exit code. presto-admin runs stuff on potentially many hosts, and
actions can fail on some hosts but not others. This change takes the approach
that in order to exit with an exit code of zero, every action taken against
every host must succeed.