Use validator result in conf
When we set conf interactively, we use the result returned by the
validator to set the conf variable. With this change we do this for
non-interactive conf as well.  This brings consistency to the conf
because we accept both strings that can be interpreted as ints and
actual ints as valid ports.

This also fixes a bug in our validation where we had not actually been
validating the port at all because we had been looking at the wrong
name.

Testing: make lint test, test_topology, test_control