Load task configuration by callback.
The presto-admin configuration should be loaded based on what the task the
user is invoking requires to run. This change does that by doing the following:

1. Unifying (somewhat) configuration between slider and standalone clusters so
that both configurations share a common base class that knows how to load
their config and whether or not the config has been loaded.

2. Adding a decorator for tasks, @requires_config, that takes a configuration
class as an argument. The decorator adds an attribute, pa_config_callback,
to the task function that main.py knows to look for and call before running
the tasks. At the time the task is run, the decorator verifies that the config
has been loaded. If the config hasn't been loaded, it's an error.

In several places tasks now explicitly have a @requires_config decorator to
express the fact that they depend on the fabric environment being set in
order to work.

There are supporting changes throughout the tests, both in the form of new
tests for main.py and the form of loading config for tests that require it.