Change check_if_valid_rpm to be run only once.
Make check_if_valid_rpm to only be run once. This required moving the call from deploy_action to install in package.py. install was annoted with @runs_once and now relies on the execute to run in the install in parallel on the cluster.

The motivation behind this change is to hopefully fix an intermittent failure during parallel calls to rpmdb on the same host. In general it makes sense to not run the same check_if_valid function more than once.