Add test suite for rpm tests
Presto-admin is not the right place to be testing the presto rpm,
but is currently the only place we can do it.  Still, we don't want
to have to run the tests for the rpm for every change in presto-admin.

Separate tests that are just for the rpm from other product tests and
add a make target for these tests. The rpm tests will not run as part
of make test-all.

Syntax:
make test-rpm

Testing: make lint test-rpm