Implement mode switching.
Mode switching serves two purposes:
1. This provides a user-friendly way to switch back and forth between managing a
YARN-slider managed cluster and a standalone cluster, which is provided with the
new 'mode' subcommand.
2. We want the user experience and interface to remain pretty consistent for
people managing standalone and YARN-slider managed presto installations. This is
done by setting up the imports in init.py such that (for example) the commands
to start and stop a cluster are exposed through the 'server' subcommand whether
you're managing a standalone or YARN-slider managed cluster.