Include pa log in test_status assertion failure messages.
test_status relies on comparing the actual output of presto-admin server status
to the expected output. Unfortunately, the status command is written in such a
way that some failures in PrestoClient are logged and otherwise silently
ignored. This results in incomplete output and an assertion failure with no
information about what the failure actually was. Since this is intermittent
and usually happens on Jenkins jobs, we don't get a chance to look at the logs
manually. The solution is to include them in the assertion failure message.