Fail build if removing docker images fails in clean-test.
We've been getting into losing streaks where some of the tests consistently
fail in automation until they get run on a new Jenkins slave. Anu noticed that
the clean-test step was failing to remove docker images; this should be an
error in automation. We should never run the product tests with images that we
created on a previous run.