Remove dummy-rpm.rpm usage from product tests
There isn't a process to keep the dummy rpm in sync with the real RPM, and
changes to the behavior of the real RPM need to be reflected in presto-admin
and its product tests. Remove the dummy RPM from the product tests to resolve
failures introduced by changes to presto-admin's behavior.