Add dummy RPM back in for upgrade product tests.
It turns out we actually need the dummy RPM for the upgrade product tests. The
upgrade tests don't depend on any of the random RPM functionality that changes
periodically, and we need *some* RPM to do an upgrade. Since downloading
another 400MB from the internet for one test is onerous, we use the dummy RPM.