docs : Add log dir details
Presto might run into issues if there is not
enough space for various logs like server.log, launcher.log
and http-requrests.log. Updated the docs to make it more
visible to the users that it is a good idea to provide
log locations that has more space and not use the default ones.

Testing: make docs, make lint