Unify cluster bringup from bare images.
test_installer and prestoadmin_installer both did a bunch of low-level
interaction with docker_cluster to bring up a bare cluster for their various
purposes. We already have code that handles bringing up a bare cluster that
base_product_test uses.

This reworks the bare cluster bringup code so that it can be used by
test_installer and prestoadmin_installer and gets them out of the business of
creating their own bare images.