Make ensure_containers_started opt-out.
Previous opt-in behavior caused problems when we change the naming of our
images. We stopped waiting for containers to fully start, and had a mess of
intermittent problems because of it. The new behavior ensures that if we change
the naming of an opted-out image, we'll see failures immediately and can
correct them.